implement order logic to project

+++ src/main/java/project/bookstore/service/impl/ShoppingCartServiceImpl.java
@@ -7,6 +7,7 @@ import project.bookstore.dto.CartItemRequestDto;
 import project.bookstore.dto.ShoppingCartResponseDto;
 import project.bookstore.dto.ShoppingCartUpdateQuantityRequestDto;
 import project.bookstore.exception.EntityNotFoundException;
+import project.bookstore.exception.OperationException;
 import project.bookstore.mapper.CartItemMapper;
 import project.bookstore.mapper.ShoppingCartMapper;
 import project.bookstore.model.CartItem;
@@ -51,8 +52,15 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
 
     @Override
     public ShoppingCartResponseDto removeBookFromShoppingCart(Long cartItemId, Long userId) {
-        cartItemRepository.deleteById(cartItemId);
-        return findByUser(userId);
+        if (isSuitableShoppingCart(cartItemId, userId)) {
+            cartItemRepository.deleteById(cartItemId);
+            return findByUser(userId);
+        }
+        throw new OperationException(
+                String.format(
+                        "It's impossible to remove book not from your own shopping cart. "
+                                + "This cart item with id: %s is situated in another shopping cart",
+                        cartItemId));
     }
 
     @Override
@@ -61,11 +69,36 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
     }
 
     @Override
+    public void clearShoppingCart(Long shoppingCartId) {
+        ShoppingCart shoppingCart = getModelById(shoppingCartId);
+        for (CartItem cartItem : shoppingCart.getCartItems()) {
+            cartItemRepository.deleteById(cartItem.getId());
+        }
+        shoppingCart.setCartItems(null);
+        shoppingCartRepository.save(shoppingCart);
+    }
+
+    @Override
+    public ShoppingCart getModelById(Long id) {
+        return shoppingCartRepository.findByUserId(id).orElseThrow(() ->
+                new EntityNotFoundException(String.format(
+                        "Can't find shop cart by user with id: %s", id)));
+    }
+
+    @Override
     public void updateCartItemQuantity(Long cartItemId,
-                                       ShoppingCartUpdateQuantityRequestDto quantityRequestDto) {
-        CartItem cartItem = getCartItemByIdFromDb(cartItemId);
-        cartItem.setQuantity(quantityRequestDto.getQuantity());
-        cartItemRepository.save(cartItem);
+                                       ShoppingCartUpdateQuantityRequestDto quantityRequestDto,
+                                       Long userId) {
+        if (isSuitableShoppingCart(cartItemId, userId)) {
+            CartItem cartItem = getCartItemByIdFromDb(cartItemId);
+            cartItem.setQuantity(quantityRequestDto.getQuantity());
+            cartItemRepository.save(cartItem);
+        }
+        throw new OperationException(
+                String.format(
+                        "It's impossible to update quantity of cart item "
+                                + "if this cart item with id: %s is belong to another (not yours)"
+                                + " shopping cart", cartItemId));
     }
 
     private CartItem getCartItemByIdFromDb(Long cartItemId) {
@@ -74,15 +107,13 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
                         String.format("Can't find cart item by id: %s", cartItemId)));
     }
 
-    private ShoppingCart getModelById(Long id) {
-        return shoppingCartRepository.findByUserId(id).orElseThrow(() ->
-                new EntityNotFoundException(String.format(
-                        "Can't find shop cart by user with id: %s", id)));
-    }
-
     private User getUserFromDb(Long userId) {
         return userRepository.findById(userId).orElseThrow(() ->
                 new EntityNotFoundException(
                         String.format("Can't find user by user id: %s", userId)));
+    }
+
+    private boolean isSuitableShoppingCart(Long cartItemId, Long userId) {
+        return getCartItemByIdFromDb(cartItemId).getShoppingCart().getId().equals(userId);
     }
 }
